requirements.txt
django
djangorestframework
drf-yasg
pandas
djangorestframework-simplejwt